Handle connection errors for the service account client
It seems that if the same client instance is open for too long, the
connection is eventually closed and the next request will fail. To
prevent "INTERNAL SERVER ERROR" from being displayed for some button
presses involving spreadsheets, the connection exception is caught and
the client or spreadsheet is refetched.