Update matches info table
To improve the user's flow, the matches info table is now more like a
"queue" that is edited by the user until the "Send Notification" button
is pressed. Until then, entering matches queries into the input will
*add* to the table (instead of replace it entirely), and the current
state of the matches info will be updated whenever a match row is
removed.