Add authentication and authorization
Authentication is done through Google (id and secret set as environment
variables), and authorization uses an Admins table. All the proper admin
endpoints are now protected with logins. The super admin is hard-coded
during the database migration, and they are able to add and remove
admins. Changing super admins (adding or removing) would require
manually editing the database; it is not possible through the frontend.

Although it was nice to abstract the navbar tabs into the views
subpackage, adding permissions for certain endpoints made it a little
more complicated to not include the logic for certain pages directly
into the Jinja template.