Allow choosing from all campaign templates
On the Notifications page, when choosing which template to use to send,
the options were previously limited to campaigns that were both in the
selected audience and the selected folder. However, campaign folders are
not restricted to a specific audience, so the number of campaigns could
have been misleading. Since the email sending will replicate the chosen
template and set the audience id anyway, there is no need to restrict
the choices to only campaigns in the selected audience. It may be
slightly confusing for users, but with proper organization of the
campaign folders, it should be fine.