Fix bug with clearing service account
When clearing the service account, `None` is passed as the data to be
set. This would then be passed to `json.dumps()`, which would obviously
output simply `"null"`. This would lead to the rest of the code thinking
that there is actually a service account set, since the info is not
the literal `null`.