Move some admin settings to a Global Settings page
The Global Settings page is only accessible by the Super Admin. The
settings here will usually never have to change, and also require some
additional ownerships of other accounts that a Super Admin would
reasonably have access to but a regular Admin may not.

Some of the routes were also changed to be more consistent, and certain
routes were moved to the `views` files that use them.