Split team codes into division and team number
Most of the UI was also changed, which results in a large commit.
However, the main reason for this change was to support a "Send to
Division" button on the Notifications page, which was not able to be
included in this commit because it requires a lot of moving around of
the Notifications page.