Select audience or tag as part of recipients list
This will also now pass the audience tag as a request argument to the
"send blast notification" endpoint, which makes it a safer operation in
case the audience tag is changed in the database before the user clicks
the button. Now the behavior will actually be what the user expects.

Also, this commit contains a lot of code cleanup and refactoring, as
well as more rigorous XSS attack prevention (there were a few places
where that could have been bad).