Use one segment for every email that is sent
A sent email on Mailchimp will use the *updated* emails from its segment
when the email is sent, which isn't necessarily when the email send is
triggered. This means that using the same segment id for multiple emails
that are all sent in quick succession will result in all the sent emails
using the email addresses for the last email, since that is email set
that most recently belongs to the segment that all the emails share.

To fix this, we will now use one segment for every email that is sent in
a batch. However, if the user tries to another email batch too close to
the previous one, they could still run into this issue.