Improve subject validation error messages
In the past, invalid placeholders such as "{match 701}" or "{701}" would
result in an error message such as "invalid character space" or "invalid
character 7", which would be confusing to users who don't understand the
placeholder system. Now, the entire placeholder will be parsed and then
checked against the valid ones, so the error message will simply say
that the entire placeholder is invalid. Hopefully this makes the error
messages easier to understand.