Add missing contacts to Mailchimp only (resolves #3)
Also changes the way the audience tag is edited on Mailchimp by simply
replacing all of its contacts by the fetched roster emails. Previously,
each individual contact was being tagged or untagged, which is very
inefficient in terms of number of API calls and time.

The changelog template was also improved so that issue numbers can now
be listed with appropriate links to the repo.